Add dokka documentation
