+++ gradle.properties
@@ -10,14 +10,14 @@ kotlin.native.ignoreDisabledTargets=true
 SONATYPE_HOST=S01
 RELEASE_SIGNING_ENABLED=true
 
-GROUP=org.kotlincrypto
+GROUP=org.kotlincrypto.random
 
 POM_INCEPTION_YEAR=2023
 
-POM_URL=https://github.com/KotlinCrypto/secure-random/
-POM_SCM_URL=https://github.com/KotlinCrypto/secure-random/
-POM_SCM_CONNECTION=scm:git:git://github.com/KotlinCrypto/secure-random.git
-POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/KotlinCrypto/secure-random.git
+POM_URL=https://github.com/KotlinCrypto/random/
+POM_SCM_URL=https://github.com/KotlinCrypto/random/
+POM_SCM_CONNECTION=scm:git:git://github.com/KotlinCrypto/random.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/KotlinCrypto/random.git
 
 POM_LICENCE_NAME=The Apache Software License, Version 2.0
 POM_LICENCE_URL=https://www.apache.org/licenses/LICENSE-2.0.txt
