+++ library/crypto-rand/api/crypto-rand.api
@@ -1,17 +1,18 @@
-public abstract class org/kotlincrypto/crypto/rand/CryptoRand {
-	public static final field Default Lorg/kotlincrypto/crypto/rand/CryptoRand$Default;
+public abstract class org/kotlincrypto/random/CryptoRand {
+	public static final field Default Lorg/kotlincrypto/random/CryptoRand$Default;
 	protected fun <init> ()V
 	public abstract fun nextBytes ([B)[B
 }
 
-public final class org/kotlincrypto/crypto/rand/CryptoRand$Default : org/kotlincrypto/crypto/rand/CryptoRand {
+public final class org/kotlincrypto/random/CryptoRand$Default : org/kotlincrypto/random/CryptoRand {
 	public fun nextBytes ([B)[B
+	public fun toString ()Ljava/lang/String;
 }
 
-public abstract interface annotation class org/kotlincrypto/crypto/rand/DelicateCryptoRandApi : java/lang/annotation/Annotation {
+public abstract interface annotation class org/kotlincrypto/random/DelicateCryptoRandApi : java/lang/annotation/Annotation {
 }
 
-public final class org/kotlincrypto/crypto/rand/RNGException : java/security/GeneralSecurityException {
+public final class org/kotlincrypto/random/RandomnessProcurementException : java/security/GeneralSecurityException {
 	public fun <init> ()V
 	public fun <init> (Ljava/lang/String;)V
 	public fun <init> (Ljava/lang/String;Ljava/lang/Throwable;)V
