+++ library/crypto-rand/src/commonMain/kotlin/org/kotlincrypto/random/internal/-CommonPlatform.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (c) 2023 Matthew Nelson
+ * Copyright (c) 2025 Matthew Nelson
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *         https://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,12 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto
+package org.kotlincrypto.random.internal
 
-@Deprecated("Deprecated in favor of CryptoRand. See https://github.com/KotlinCrypto/random")
-public class SecRandomCopyException: RuntimeException {
-    public constructor(): super()
-    public constructor(message: String?): super(message)
-    public constructor(message: String?, cause: Throwable?): super(message, cause)
-    public constructor(cause: Throwable?): super(cause)
-}
+import org.kotlincrypto.random.RandomnessProcurementException
+
+@Throws(RandomnessProcurementException::class)
+internal expect fun ByteArray.cryptoRandFill()
