--- library/crypto-rand/src/darwinMain/kotlin/org/kotlincrypto/crypto/rand/internal/DarwinPlatform.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2025 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package org.kotlincrypto.crypto.rand.internal
-
-import kotlinx.cinterop.*
-import platform.Security.SecRandomCopyBytes
-import platform.Security.kSecRandomDefault
-import org.kotlincrypto.crypto.rand.RNGException
-
-@Throws(RNGException::class)
-internal actual fun ByteArray.cryptoRandFill() {
-    @OptIn(ExperimentalForeignApi::class)
-    val status: Int = usePinned { pinned ->
-        // kSecRandomDefault is synonymous to NULL
-        SecRandomCopyBytes(kSecRandomDefault, size.toUInt().convert(), pinned.addressOf(0))
-    }
-
-    if (status == 0) return
-    throw RNGException("Failed to obtain bytes from [SecRandomCopyBytes]. Code: $status")
-}
