+++ library/secure-random/gradle.properties
@@ -14,3 +14,5 @@
 POM_ARTIFACT_ID=secure-random
 POM_NAME=KotlinCrypto SecureRandom
 POM_DESCRIPTION=KotlinCrypto library for obtaining cryptographically secure random data from the system
+
+GROUP=org.kotlincrypto
